bookmarks: updated bookmarks logic for css styles & html labels & logic to allow multiple files imports + refactoring import logic to move it to renderer
