+++ app/colors/default.css
@@ -420,9 +420,10 @@ body:not(#app) ::placeholder {color: var(--placeholder-text);}
 #historypage .hist-entry .favicon {float: none;margin: 0 .3em .1em 0;cursor: auto;}
 #historypage .hist-entry.marked {background: var(--special-page-element-border);}
 /* bookmarks */
-#bookmarks button{float: right;}
-#bookmarks .bookmark,.folder{margin-top: 5px; padding: 5px 0 5px 5px;}
-#bookmarks .keywords{margin-top:  7px; padding: 5px; font-size: 10px; font-style: italic; text-transform: lowercase;}
+#bookmarkspage button {float: right;}
+#bookmarkspage .bookmark, #bookmarkspage .folder {margin-top: .3em; padding: .3em 0 .3em .3em;}
+#bookmarkspage .keywords {margin-top:  .4em; padding: .3em; font-size: .6em; font-style: italic; text-transform: lowercase;}
+#bookmarkspage ul {margin-top: .9em;}
 /* newtab */
 #newtabpage {display: none;flex-direction: column;height: 100vh;}
 #newtabpage main {margin: auto;max-width: 80vw;}
