+++ app/pages/bookmarks.html
@@ -24,10 +24,8 @@
     <link rel="shortcut icon" type="image/png" href="../img/vieb.svg">
 </head>
 <body class="specialpage" id="bookmarkspage">
-    <div class="specialpage-controls">
-        <h1>Bookmarks</h1>
-        <button id="import-bookmarks-button" class="button">Import bookmarks</button>
-    </div>
+    <h1>Bookmarks</h1>
+    <button id="import-bookmarks-button">Import bookmarks</button>
     <div id="bookmarks"></div>
 </body>
 </html>
