+++ app/pages/help.html
@@ -666,10 +666,10 @@
         With this command you can delete any previously stored <a class="pointer" href="#pointer">pointer</a> positions by key, domain or url. Whether it uses the domain or the url is controlled by <a href="#pointerposlocalid">pointerposlocalid</a>. Without a confirm override "!", it will always require a key to delete, which is by default only for the local ones at the current domain/url or the global ones. If you pass a domain/url, you can delete the key for a different domain/url, where "*" will match all of them including global ones. If you do supply the "!" confirm override, you can delete any key in one go, by default for the current domain/url, but you can also supply one. In case you supply "*", "local" or "global", it will delete ALL, local or global positions respectively.
         <h2 id="bookmarks">Bookmarks</h2>
         With Bookmarks, you can quickly store page urls for later use. Besides commands, there are also settings to control the way they are stored and loaded.
-        <h3 id=":bm">:bm</h3>
+        <h3 id=":bookmarks">:bookmarks</h3>
         <ul>
-            <li><span class="command-block">:bm</span> - Open the bookmarks page in a new tab unless it should replace the current based on <a href="#replacespecial">replacespecial</a></li>
-            <li><span class="command-block">:bm!</span> - Open the bookmarks page in a new tab regardless of settings</li>
+            <li><span class="command-block">:bookmarks</span> - Open the bookmarks page in a new tab unless it should replace the current based on <a href="#replacespecial">replacespecial</a></li>
+            <li><span class="command-block">:bookmarks!</span> - Open the bookmarks page in a new tab regardless of settings</li>
         </ul>
         This command will open the bookmarks page, located at "vieb://bookmarks". On this page, you will find a list of all your bookmarks, and a button to import bookmarks from a Chrome bookmarks HTML file.
         <h3 id=":bmadd">:bmadd</h3>
