+++ app/translations/jp.json
@@ -35,7 +35,7 @@
         "deleted": "クエリに一致した$1個のブックマークを削除しました。",
         "exists": "名前が$1のブックマークは既に存在します。",
         "folder": {
-            "deleted": "クエリに一致したフォルダを削除しました"
+            "deleted": "クエリに一致したフォルダを削除しました。"
         },
         "hex": {
             "invalid": "無効なカラーヘックス $1"
