+++ app/translations/nl.json
@@ -55,7 +55,7 @@
         },
 
         "loaded": "$1 bladwijzers geladen die overeenkwamen met de zoekopdracht.",
-        "noname": "Geen naam opgegeven voor de bladwijzer",
+        "noname": "Geen naam opgegeven voor de bladwijzer.",
         "notfound": "Kon bladwijzer niet vinden."
     },
     "commands": {
